4. Учебный проект: время разбивать камни
