+++ src/view/board-tasks.js
@@ -1,0 +1,5 @@
+export const createBoardTasksTemplate = () => {
+  return (
+    `<div class="board__tasks"></div>`
+  );
+};
