+++ src/view/board.js
@@ -1,0 +1,5 @@
+export const createBoardContainerTemplate = () => {
+  return (
+    `<section class="board container"></section>`
+  );
+};
